split tests by api method
